+++ src/Installer.php
@@ -2,6 +2,7 @@
 
 namespace luya\composer;
 
+use Composer\Config;
 use Composer\Installer\LibraryInstaller;
 use Composer\Repository\InstalledRepositoryInterface;
 use Composer\Package\PackageInterface;
@@ -18,46 +19,70 @@ class Installer extends LibraryInstaller
     
     const LUYA_FILE = 'luyadev/installer.php';
     
-    private $_relativeVendorDir;
-    
+    /**
+     * {@inheritDoc}
+     */
     public function supports($packageType)
     {
         return $packageType == 'luya-core' || $packageType == 'luya-extension' || $packageType == 'luya-module';
     }
     
+    /**
+     * {@inheritDoc}
+     */
     public function install(InstalledRepositoryInterface $repo, PackageInterface $package)
     {
         // install the package the normal composer way
         parent::install($repo, $package);
-        
         $this->addPackage($package);
     }
     
+    /**
+     * {@inheritDoc}
+     */
     public function update(InstalledRepositoryInterface $repo, PackageInterface $initial, PackageInterface $target)
     {
         parent::update($repo, $initial, $target);
-        
         $this->removePackage($initial);
         $this->addPackage($target);
     }
     
+    /**
+     * {@inheritDoc}
+     */
     public function uninstall(InstalledRepositoryInterface $repo, PackageInterface $package)
     {
         parent::uninstall($repo, $package);
-        
         $this->removePackage($package);
     }
     
+    /**
+     * Add a package to installer
+     *
+     * @param PackageInterface $package
+     * @return void
+     */
     protected function addPackage(PackageInterface $package)
     {
         $this->writeInstaller($this->addConfig($package));
     }
     
+    /**
+     * Remove a package from installer
+     *
+     * @param PackageInterface $package
+     * @return void
+     */
     protected function removePackage(PackageInterface $package)
     {
         $this->writeInstaller($this->removeConfig($package));
     }
     
+    /**
+     * Get the installer array
+     *
+     * @return array
+     */
     protected function getInstallers()
     {
         $file = $this->vendorDir . DIRECTORY_SEPARATOR . self::LUYA_FILE;
@@ -75,19 +100,37 @@ class Installer extends LibraryInstaller
         return $data;
     }
     
+    /**
+     * Ensure a config for a package.
+     *
+     * @param PackageInterface $package
+     * @param array $config
+     * @return void
+     */
     protected function ensureConfig(PackageInterface $package, array $config)
     {
+        // generate the package folder, which is actually the name but with os based directory seperator
+        $packageFolder = str_replace("/", DIRECTORY_SEPARATOR, $package->getPrettyName());
+        
         $packageConfig = [
-            'package' => ['name' => $package->getName(), 'prettyName' => $package->getPrettyName(), 'version' => $package->getVersion()],
+            'package' => [
+                'name' => $package->getName(),
+                'prettyName' => $package->getPrettyName(),
+                'version' => $package->getVersion(),
+                'targetDir' => $package->getTargetDir(),
+                'installSource' => $package->getInstallationSource(),
+                'sourceUrl' => $package->getSourceUrl(),
+                'packageFolder' => $packageFolder,
+            ],
             'blocks' => [],
-            'bootstrap' => (isset($config['bootstrap'])) ? $config['bootstrap'] : [],
+            'bootstrap' => (isset($config['bootstrap'])) ? ComposerHelper::parseDirectorySeperator($config['bootstrap']) : [],
             'themes' => [],
         ];
         
         $blocks = (isset($config['blocks'])) ? $config['blocks'] : [];
     
         foreach ($blocks as $blockFolder) {
-            $packageConfig['blocks'][] = $this->getRelativeVendorDir() . DIRECTORY_SEPARATOR . $package->getPrettyName() . DIRECTORY_SEPARATOR . ltrim($blockFolder, '/');
+            $packageConfig['blocks'][] = $this->getRelativeVendorDir() . DIRECTORY_SEPARATOR . $packageFolder . DIRECTORY_SEPARATOR . ComposerHelper::parseDirectorySeperator(ltrim($blockFolder, DIRECTORY_SEPARATOR));
         }
     
         $themes = (isset($config['themes'])) ? $config['themes'] : [];
@@ -98,7 +141,13 @@ class Installer extends LibraryInstaller
         
         return $packageConfig;
     }
-        
+     
+    /**
+     * Remove a package from the config
+     *
+     * @param PackageInterface $package
+     * @return array
+     */
     protected function removeConfig(PackageInterface $package)
     {
         $data = $this->getInstallers();
@@ -110,6 +159,12 @@ class Installer extends LibraryInstaller
         return $data;
     }
     
+    /**
+     * Get the LUYA extra binary data.
+     *
+     * @param PackageInterface $package
+     * @return array
+     */
     protected function getPackageExtraData(PackageInterface $package)
     {
         if (empty($package->getExtra())) {
@@ -119,6 +174,12 @@ class Installer extends LibraryInstaller
         return isset($package->getExtra()[self::LUYA_EXTRA]) ? $package->getExtra()[self::LUYA_EXTRA] : [];
     }
     
+    /**
+     * Add a package to the config
+     *
+     * @param PackageInterface $package
+     * @return array
+     */
     protected function addConfig(PackageInterface $package)
     {
         $data = $this->getInstallers();
@@ -127,6 +188,12 @@ class Installer extends LibraryInstaller
         return $data;
     }
     
+    /**
+     * Write the installer.php file in vendor folder
+     *
+     * @param array $data
+     * @return void
+     */
     protected function writeInstaller(array $data)
     {
         $file = $this->vendorDir . DIRECTORY_SEPARATOR . self::LUYA_FILE;
@@ -145,6 +212,8 @@ class Installer extends LibraryInstaller
             @opcache_invalidate($file, true);
         }
     }
+
+    private $_relativeVendorDir;
     
     /**
      * Read the relative vendor-dir from composer config.
@@ -155,11 +224,9 @@ class Installer extends LibraryInstaller
     public function getRelativeVendorDir()
     {
         if ($this->_relativeVendorDir === null) {
-            $this->_relativeVendorDir = rtrim($this->composer->getConfig()->get('vendor-dir', \Composer\Config::RELATIVE_PATHS), '/');
+            $this->_relativeVendorDir = rtrim($this->composer->getConfig()->get('vendor-dir', Config::RELATIVE_PATHS), DIRECTORY_SEPARATOR);
         }
         
         return $this->_relativeVendorDir;
     }
-    
-    
 }